requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-extensions==3.2.3
django-js-asset==2.2.0
idna==3.7
msgpack-python==0.5.6
mysqlclient==2.2.4
pillow==10.3.0
razorpay==1.4.2
redis==2.10.6
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
